+++ src/test/java/org/sopt/app/application/FriendServiceTest.java
@@ -12,7 +12,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.sopt.app.application.poke.FriendService;
-import org.sopt.app.interfaces.postgres.FriendRepository;
+import org.sopt.app.interfaces.postgres.friend.FriendRepository;
 
 @ExtendWith(MockitoExtension.class)
 class FriendServiceTest {
