+++ src/test/java/org/sopt/app/interfaces/postgres/FriendRepositoryTest.java
@@ -8,11 +8,11 @@ import org.junit.jupiter.api.Test;
 import org.sopt.app.common.config.QuerydslConfiguration;
 import org.sopt.app.domain.entity.Friend;
 import org.sopt.app.domain.entity.User;
+import org.sopt.app.interfaces.postgres.friend.FriendRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 
 @DataJpaTest
